Add Start and Stop methods to CameraController, update system state
Added OnStart and OnStop virtual methods to the CameraController class. This allows for proper initialization and cleanup actions to be performed when the camera system starts and stops, respectively.
Fixed a bug in the GameOfLifeSystem where entity pools were not being cleared on system start. This was causing previously deleted entities to persist across system restarts.